ifnot

+++ fedora/setup_scripts/install_terraform.sh
@@ -5,10 +5,14 @@ if [[ $EUID -ne 0 ]]; then
    exit 1
 fi
 
+TF_VERSION=0.14.5
 SUDO_USER_HOME="$(eval echo "~$SUDO_USER")"
 echo "Installing terraform"
 mkdir -p $SUDO_USER_HOME/bin
 cd $SUDO_USER_HOME/bin
-wget https://releases.hashicorp.com/terraform/0.14.5/terraform_0.14.5_linux_amd64.zip
-unzip terraform_0.14.5_linux_amd64.zip
-rm terraform*zip
+
+if [ ! -d "terraform" ]; then
+    wget https://releases.hashicorp.com/terraform/$TF_VERSION/terraform_$TF_VERSION_linux_amd64.zip
+    unzip terraform_$TF_VERSION_linux_amd64.zip
+    rm terraform*zip
+fi
